Cast game result as integer to print properly

+++ src/tools/transform.cpp
@@ -445,7 +445,7 @@ namespace Stockfish::Tools
 			    sync_cout <<
 				pos.game_ply() << "," <<
 				pos.fen() << "," <<
-				ps.score << "," << ps.game_result << "," <<
+				ps.score << "," << (int)ps.game_result << "," <<
  			        UCI::move((Stockfish::Move)ps.move, false) <<
 				sync_endl;
 		    } else {
@@ -458,7 +458,7 @@ namespace Stockfish::Tools
 			    sync_cout <<
 				pos.game_ply() << "," <<
 				pos.fen() << "," <<
-				ps.score << "," << ps.game_result << "," <<
+				ps.score << "," << (int)ps.game_result << "," <<
  			        UCI::move((Stockfish::Move)ps.move, false) << "," <<
 				"d6 pv2," <<
 				UCI::move((Stockfish::Move)best_move, false) << "," <<
@@ -471,7 +471,7 @@ namespace Stockfish::Tools
 			    sync_cout <<
 				pos.game_ply() << "," <<
 				pos.fen() << "," <<
-				ps.score << "," << ps.game_result << "," <<
+				ps.score << "," << (int)ps.game_result << "," <<
  			        UCI::move((Stockfish::Move)ps.move, false) << "," <<
 				"d6 pv2," <<
 				UCI::move((Stockfish::Move)best_move, false) << "," << m1_score <<
